supplicant: only signal ScanDone if the scan really is done (or failed)
It's counter-intuitive to signal a ScanDone when the request to Scan succeeds
with the supplicant, because we'll get one ScanDone very early one when
the scan begins, which will trigger the usual mechanisms, and again once
the scan really is finished.

It does still make sense however to signal when the request fails.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>